Agriculture data from Pakistan Social and Living Standard Measures 2015

Agriculture data from Pakistan Social and Living Standard Measures 2015.

data(Agriculture)

Format

A data.table and data.frame with 3324 observations of 97 variables.

hhcode

Household 10 digits code.

Province

Province of Pakistan

Region

Region of Pakistan (Rural/Urban)

PSU

primary sampling unit 8 digits code

a101

Own Land Status (Yes/No)

a102

Owned Land (Acres)

a103

Rent Out Land Status (Yes/No)

a104

Rented Out Land (Acres)

a105

Rent Received (Rs)

a106

Rented Land Status(Yes/No)

a107

Rent Paid (Rs)

a108a

Amount Received by Sold Agriculture Land (Rs.)

a108b

Value of land Received by Gift/Inheritence (Rs.)

a108c

Amount Paid for Purchasing Agriculture Land (Rs.)

a108d

Value of land Given-away/Lost (Rs.)

a109a

Owned operational land (Acres)

a109b

Rented operational land (Acres)

a109c

Share crops basis operational land (Acres)

a109d

Any other basis operational land (Acres)

a110

Total operational land (Acres)

a111a

Irrigated operational cultivated land (Acres)

a111b

Non-irrigated operational cultivated land (Acres)

a111c

Uncultivated land (Acres)

a112

Total cultivated land (Acres)

a113

Cost of owned land improvements (Rs.)

a114

Cost of improvements of other than land (Rs.)

a121a

Wheat harvested per kg

a121b

Wheat harvested per 40 kg

a121c

Value of total Wheat production (Rs.)

a121d

Value of wheat given to landlord (Rs.)

a121e

Value of wheat kept by the household (Rs.)

a122a

Cotton harvested per kg

a122b

Cotton harvested per 40 kg

a122c

Value of total Cotton production (Rs.)

a122d

Value of cotton given to landlord (Rs.)

a122e

Value of cotton kept by the household (Rs.)

a123a

Sugarcane harvested per kg

a123b

Sugarcane harvested per 40 kg

a123c

Value of total sugarcane production (Rs.)

a123d

Value of sugarcane given to landlord (Rs.)

a123e

Value of sugarcane kept by the household (Rs.)

a124a

Rice harvested per kg

a124b

Rice harvested per 40 kg

a124c

Value of total rice production (Rs.)

a124d

Value of rice given to landlord (Rs.)

a124e

Value of rice kept by the household (Rs.)

a125a

Maize harvested per kg

a125b

Maize harvested per 40 kg

a125c

Value of total maize production (Rs.)

a125d

Value of maize given to landlord (Rs.)

a125e

Value of maize kept by the household (Rs.)

a126a

Pulses harvested per kg

a126b

Pulses harvested per 40 kg

a126c

Value of total pulses production (Rs.)

a126d

Value of pulses given to landlord (Rs.)

a126e

Value of pulses kept by the household (Rs.)

a127a

Fruit harvested per kg

a127b

Fruit harvested per 40 kg

a127c

Value of total fruit production (Rs.)

a127d

Value of fruit given to landlord (Rs.)

a127e

Value of fruit kept by the household (Rs.)

a128a

Vegetables harvested per kg

a128b

Vegetables harvested per 40 kg

a128c

Value of total vegetables production (Rs.)

a128d

Value of vegetables given to landlord (Rs.)

a128e

Value of vegetables kept by the household (Rs.)

a129a

Fodder harvested per kg

a129b

Fodder harvested per 40 kg

a129c

Value of total fodder production (Rs.)

a129d

Value of fodder given to landlord (Rs.)

a129e

Value of fodder kept by the household (Rs.)

a130a

Other crop harvested per kg

a130b

Other crop harvested per 40 kg

a130c

Value of total other crop production (Rs.)

a130d

Value of other crop given to landlord (Rs.)

a130e

Value of other crop kept by the household (Rs.)

a131a

Bi-products crops harvested per kg

a131b

Bi-products crops harvested per 40 kg

a131c

Value of total bi-product crop production (Rs.)

a131d

Value of bi-product crop given to landlord (Rs.)

a131e

Value of bi-product crop kept by the household (Rs.)

a135c

Total crops harvested per kg

a135d

Total crops harvested per 40 kg

a135e

Value of total crops production (Rs.)

a135f

Value of total crops given to landlord (Rs.)

a135g

Value of total crops kept by the household (Rs.)

a136

Cost on seeds/plants (Rs.)

a137

Cost on fertilizer (Rs.)

a138

Cost on persticides (Rs.)

a139

Cost on water/electricity/fuel (Rs.)

a140

All types of taxes paid (Rs.)

a141

Freight/transportation/commission/insurance/storage charges (Rs.)

a142

Permanent labour charges (Rs.)

a143

Casual labour charges (Rs.)

a144

Rent of equipment/animal charges (Rs.)

a145

Other expenses (Rs.)

a150

Total expenses (Rs.)

References

Pakistan Bureau of Statistics, Micro data (http://www.pbs.gov.pk/content/microdata).

See also

Examples

 # library(PSLM2015)
 # data("Agriculture")
 # library(dplyr)
 # Agriculture %>%
 #   group_by(Province, Region) %>%
 #   summarise(TotalOperationalLand = sum(a110, na.rm = TRUE))
 # library(ggplot2)
 # ggplot(data = Agriculture, mapping = aes(x = Province, y = a110)) + 
 #   geom_col() +
 #   labs(y = "Total Operational Land") +
 #   facet_grid(. ~ Region)
 # 
 #   # Merging two data files
 #   data("Employment")
 #   data("Agriculture")
 #    ab <- Employment %>% 
 #    filter(s1bq06 %in% 
 #    c("Own cultivator","Share cropper", "Contract cultivator")
 #    |s1bq14 %in% c("Own cultivator","Share cropper", "Contract cultivator"))
 #   
 #   EmpAgri <- ab %>% left_join(Agriculture, by = c("hhcode", "Province", "Region", "PSU"))
 #   str(EmpAgri)