Education data from Pakistan Social and Living Standards Measurement 2015-16

Education data from Pakistan Social and Living Standards Measurement 2015-16.

data(Education)

Format

A data.table and data.frame with 141828 observations of 22 variables.

hhcode

Household 10 digits code.

Province

Province of Pakistan

Region

Region of Pakistan (Rural/Urban)

PSU

primary sampling unit 8 digits code

idc

Identity code of household member

s2ac01

Can read with understanding

s2ac02

Can Write with understanding

s2ac03

Can solve arithmatic questions

s2ac04

Attended any educational institution

s2ac05

Highest level of education passed

s2ac06

Currently attending educational institution

s2ac07

Currently studying class

s2ac08

Type of currently attending institution

s2ac9a

Last year expenditure on school Fees/Admission/Registration/Funds/Donations?

s2ac9b

Last year expenditure on school Uniform?

s2ac9c

Last year expenditure on school Books/stationery items?

s2ac9d

Last year expenditure on school Examination Fee?

s2ac9e

Last year expenditure on Private Tuition?

s2ac9f

Last year expenditure on school transportation?

s2ac9g

Last year expenditure on school hostel expenses?

s2ac9h

Last year expenditure on school other expenses?

s2ac9i

Total expenditure on schooling

References

Pakistan Bureau of Statistics, Micro data (http://www.pbs.gov.pk/content/microdata).

See also

Examples

 # library(PSLM2015)
 # library(dplyr)
 # data("Education")
 # TotalP <- Education %>% group_by(Province, Region) %>%
 #   summarise(TotalPersons = n())
 #   
 # literacy <- Education %>% filter(s2ac01 == "yes" & s2ac02 == "yes" & s2ac03 == "yes")
 #   literateP <- literacy %>%
 #   group_by(Province, Region) %>%
 #   summarise(literatePersons = n())
 #   literacyR <- TotalP %>% left_join(literateP, by = c("Province", "Region"))
 #   literacyRate <- mutate(literacyR, Rate = literatePersons/TotalPersons*100)
 # library(ggplot2)
 # ggplot(data = literacyRate, mapping = aes(x = Province, y = Rate)) +
 #   geom_col() +
 #   facet_grid(. ~ Region)
 # 
 # # Merging two data files
 # 
 #   data("Employment")
 #   data("Education")
 #   income <- Employment %>% rowwise() %>%
 #     mutate(TotalIncome = sum((s1bq08*s1bq09),s1bq10,s1bq15,s1bq17,s1bq19,s1bq21, na.rm = TRUE))
 #   ab <- income %>% select(hhcode, idc, TotalIncome)
 #   EduEmp <- Education %>% left_join(ab, by = c("hhcode", "idc"))
 #   str(EduEmp)