Employment and income data from Pakistan Social and Living Standards Measurement 2015-16

Employment and income data from Pakistan Social and Living Standards Measurement 2015-16.

data(Employment)

Format

A data.table and data.frame with 115910 observations of 27 variables.

hhcode

Household 10 digits code.

Province

Province of Pakistan

Region

Region of Pakistan (Rural/Urban)

PSU

primary sampling unit 8 digits code

idc

Identity code of household member

s1bq01

Last month working status

s1bq02

Number of worked days in last month

s1bq03

Employment/business/economic activity status

s1bq04

Occupation

s1bq05

Industry

s1bq06

Type of economic activity

s1bq07

Income reporting (Monthly/Anually)

s1bq08

Last month cash income (Rs.)

s1bq09

Number of months worked in last year

s1bq10

Last year cash income (Rs.)

s1bq11

Part time working status

s1bq12

Part time occupation

s1bq13

Part time working industry

s1bq14

Part time economic activity type

s1bq15

Last year part time cash income (Rs.)

s1bq16

Any other work done for pay/profit in last year (Yes/No)

s1bq17

Last year cash income from other work (Rs.)

s1bq18

Sold status of in kind wages (Yes/No)

s1bq19

Last year income by selling in-kind wages (Rs.)

s1bq20

Pension or other financial benefits in last year (Yes/No)

s1bq21

Last year income from pension/other financial benefits (Rs.)

s1bq22

Income used to pay expences of household (Rs.)

References

Pakistan Bureau of Statistics, Micro data (http://www.pbs.gov.pk/content/microdata).

See also

Examples

 # library(PSLM2015)
 # data("Employment")
 # library(dplyr)
 # x2<- distinct(Employment, hhcode, .keep_all = TRUE)
 # TotalHH<- x2 %>% group_by(Province, Region) %>%
 #   summarise(TotalHH = n())
 # income<- Employment %>% rowwise() %>%
 #   mutate(TotalIncome = sum((s1bq08*s1bq09),s1bq10,s1bq15,s1bq17,s1bq19,s1bq21, na.rm = TRUE))
 # IncomeR <- income %>%
 #   group_by(Province, Region) %>%
 #   summarise(TotalIncome = sum(as.numeric(TotalIncome)))
 # IncomeR2 <- TotalHH %>% left_join(IncomeR, by = c("Province", "Region"))
 # IncomeRate <- IncomeR2 %>% mutate(AverageHHIncome = TotalIncome/TotalHH)
 # 
 # library(ggplot2)
 # ggplot(data = IncomeRate, mapping = aes(x = Province, y = AverageHHIncome)) +
 #   geom_col() +
 #   facet_grid(. ~ Region)
 # 
 # # Merging two data files
 # 
 #   data("Employment")
 #   data("Education")
 #   income <- Employment %>% rowwise() %>%
 #     mutate(TotalIncome = sum((s1bq08*s1bq09),s1bq10,s1bq15,s1bq17,s1bq19,s1bq21, na.rm = TRUE))
 # ab <- select(income, hhcode, idc, TotalIncome)
 # EduEmp<-Education %>% left_join(ab, by = c("hhcode", "idc"))
 # str(EduEmp)