Housing data from Pakistan Social and Living Standards Measurement 2015-16

Housing data from Pakistan Social and Living Standards Measurement 2015-16.

data(Housing)

Format

A data.table and data.frame with 24238 observations of 36 variables.

hhcode

Household 10 digits code.

Province

Province of Pakistan

Region

Region of Pakistan (Rural/Urban)

PSU

primary sampling unit 8 digits code

S3aq01

Dwelling type

S3aq02

Occupancy status

S3aq03

Estimated rent of the house (Rs.)

S3aq04

Number of rooms in household

S3aq05A

Electricity facility

S3aq05B

Gas facility

S3aq06

Source of drinking water

S3aq07

Water availability (hours)

S3aq08

Water system installed by

S3aq09

Water system look-after by

S3aq10

Distance of source of drinking water (Km.)

S3aq11

Time consumption in fetching drinking water (Minutes)

S3aq12

Water payment status (Yes/No)

S3aq13

One month payment for water (Rs.)

S3aq14

Willingness to improve water supply system (Yes/No)

S3aq15

Toilet used by household

S3aq16

Defecation/urination place

S3aq17

Is your house connected with drainage/swerage system?

S3aq18A

Garbage collected by

S3aq18B

Garbage collected in neighbourhod by

S3aq19A

Monthly expenditure on household's garbage collection

S3aq19B

Monthly expenditure on neighbourhood's garbage collection

S3aq20A

Internet facility in household (Yes/No)

S3aq20B

Broad band facility in household (Yes/No)

S3aq20C

Mobile facility in household (Yes/No)

S3aq20D

Landline facility in household (Yes/No)

S3aq20E

Desktop computer facility in household (Yes/No)

S3aq20F

Laptop facility in household (Yes/No)

S3aq20G

Tablet facility in household (Yes/No)

S3aq20H

I-pad facility in household (Yes/No)

S3aq21A

Type of internet services

S3aq21B

Type of internet services for broadband

References

Pakistan Bureau of Statistics, Micro data (http://www.pbs.gov.pk/content/microdata).

See also

Examples

 # library(PSLM2015)
 # data("Housing")
 # library(dplyr)
 # AvgRooms <- Housing %>%
 #   group_by(Province, Region) %>%
 #   summarise(AvgRooms = mean(S3aq04, na.rm = TRUE))
 # 
 # library(ggplot2)
 # ggplot(data = AvgRooms , mapping = aes(x = Province, y = AvgRooms)) +
 #   geom_col() +
 #   facet_grid(. ~ Region)
 # 
 # # Merging two data files
 # 
 #   data("Employment")
 #   data("Housing")
 #   HeadHH <- HHRoster %>% filter(s1aq02 == "Head")
 #   EmpHous <- HeadHH %>% left_join(Housing, by = c("hhcode"))
 #   str(EmpHous)